# swagger.yaml
swagger: "2.0"
info:
  description: "Authors Haven - A social platform for the creative at heart."
  version: "1.0.0"
  title: "Authors Haven"
  license:
    name: "MIT"
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/api/v1"
...
schemes:
  - "https"
  - "http"
paths:
  /users:
    post:
      tags:
      - "users"
      summary: "Create user"
      description: "Creates a new user in the application"
      operationId: "createUser"
      produces:
      - "application/json"
      - "application/xml"
      parameters:
      - in: "formData"
        name: "username"
        description: "The unique username"
        required: true
        type: "string"
        format: "text"
        minLength: 4
        maxLength: 100
      - in: "formData"
        name: "email"
        description: "The unique user email"
        required: true
        type: "string"
        format: "email"
        minLength: 5
        maxLength: 255
      - in: "formData"
        name: "password"
        description: "The user password"
        required: true
        type: "string"
        format: "password"
        minLength: 5
        maxLength: 255
      responses:
        201:
          description: "User created successfully"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid email/password supplied"
    get:
      tags:
      - "users"
      summary: "Users with profiles"
      description: "Returns all users and their corresponding profiles"
      operationId: "getUsers"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "fail"
      security:
      - api_key: []
    put:
      tags:
      - "profiles"
      summary: "Update User Profile"
      description: "Returns the Updated User Details"
      operationId: "updateProfile"
      produces:
      - "application/json"
      parameters:
      - in: "formData"
        name: firstname
        required: false
        type: string
        description: "first name"
      - in: "formData"
        name: lastname
        required: false
        type: string
        description: "last name"
      - in: "formData"
        name: bio
        required: false
        type: text
        description: "bio"
      - in: "formData"
        name: avatar
        required: false
        type: string
        description: "Avatar URL"
      - in: "formData"
        name: location
        required: false
        type: string
        description: "Location"
      - in: "formData"
        name: phone
        required: false
        type: integer
        description: 'Phone Number'
        minLength: 11
        maxLength: 14
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        400:
          description: "Invalid token supplied: format Bearer <token>"
        401:
          description: "Invalid Token Provided"
      security:
      - api_key: []
...
  /profiles/{username}:
    get:
      tags:
      - "profiles"
      summary: "User Profile details"
      description: "Returns the profile details of a user"
      operationId: "getProfile"
      produces:
      - "application/json"
      parameters:
        - in: path
          name: username
          required: false
          type: string
          description: "User's username"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/ApiResponse"
        404:
          description: "fail"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  ApiResponse:
    type: "object"
    properties:
      status:
        type: "integer"
        format: "int32"
      data:
        type: "object"
        items:
          type: "object"
...